# #451 optimize array handling