… Jint/Native/Object/ObjectInstance.cs
… Jint/Native/String/StringPrototype.cs
… Jint/Options.cs
… Jint/Pooling/CompletionPool.cs
… Jint/Runtime/Completion.cs
… Jint/Runtime/Environments/LexicalEnvironment.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/StatementInterpreter.cs
… Jint/Runtime/TypeConverter.cs