* use enum for CompletionType
* use fields for frequently accessed options
* move push and map array instance to allow more optimizations
* add generic FindWithCallback
* add find and findIndex implementations, `some` can use it
* change Completion to struct instead of pooled class - no big perf difference or for the better - makes code a lot simpler

## ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Slice|100|525.0 us|161.1328|660.16 KB|
| **New** |	|	| **435.6 us (-17%)** | **161.1328 (0%)** | **660.16 KB (0%)** |
| Old |Concat|100|603.8 us|175.7813|720.31 KB|
| **New** |	|	| **474.1 us (-21%)** | **175.7813 (0%)** | **720.31 KB (0%)** |
| Old |Unshift|100|20,794.4 us|3562.5000|14672.66 KB|
| **New** |	|	| **17,848.0 us (-14%)** | **3562.5000 (0%)** | **14672.66 KB (0%)** |
| Old |Push|100|13,049.5 us|515.6250|2134.38 KB|
| **New** |	|	| **10,870.3 us (-17%)** | **343.7500 (-33%)** | **1438.28 KB (-33%)** |
| Old |Index|100|12,720.8 us|390.6250|1637.5 KB|
| **New** |	|	| **12,397.2 us (-3%)** | **390.6250 (0%)** | **1637.5 KB (0%)** |
| Old |Map|100|3,654.8 us|765.6250|3151.56 KB|
| **New** |	|	| **3,313.6 us (-9%)** | **765.6250 (0%)** | **3149.22 KB (0%)** |
| Old |Apply|100|698.1 us|190.4297|782.81 KB|
| **New** |	|	| **584.8 us (-16%)** | **188.4766 (-1%)** | **774.22 KB (-1%)** |
| Old |JsonStringifyParse|100|4,679.7 us|1273.4375|5233.59 KB|
| **New** |	|	| **4,529.2 us (-3%)** | **1273.4375 (0%)** | **5225 KB (0%)** |


## ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|688.4 ms|60000.0000|8312.5000|256.64 MB|
| **New** |	|	| **605.4 ms (-12%)** | **60000.0000 (0%)** | **2937.5000 (-65%)** | **245.9 MB (-4%)** |


## DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|80.69 ms|1437.5000|500.0000|-|7749.97 KB|
| **New** |	|	| **78.39 ms (-3%)** | **1500.0000 (+4%)** | **500.0000 (0%)** | **-** | **7747.67 KB (0%)** |
| Old |Run|dromaeo-core-eval|19.20 ms|62.5000|-|-|299.02 KB|
| **New** |	|	| **18.05 ms (-6%)** | **62.5000 (0%)** | **-** | **-** | **299.02 KB (0%)** |
| Old |Run|dromaeo-object-array|183.06 ms|43312.5000|2000.0000|1000.0000|181953.83 KB|
| **New** |	|	| **163.65 ms (-11%)** | **42562.5000 (-2%)** | **2000.0000 (0%)** | **1000.0000 (0%)** | **178313.77 KB (-2%)** |
| Old |Run|dromaeo-object-regexp|1,083.92 ms|62312.5000|33187.5000|22875.0000|468378.2 KB|
| **New** |	|	| **897.01 ms (-17%)** | **53187.5000 (-15%)** | **32375.0000 (-2%)** | **21812.5000 (-5%)** | **414189.39 KB (-12%)** |
| Old |Run|dromaeo-object-string|1,249.79 ms|221312.5000|177125.0000|173812.5000|1403605.85 KB|
| **New** |	|	| **1,224.70 ms (-2%)** | **221750.0000 (0%)** | **178562.5000 (+1%)** | **175875.0000 (+1%)** | **1392662.88 KB (-1%)** |
| Old |Run|dromaeo-string-base64|187.88 ms|6062.5000|437.5000|-|26624.45 KB|
| **New** |	|	| **178.84 ms (-5%)** | **6125.0000 (+1%)** | **312.5000 (-29%)** | **-** | **26272.45 KB (-1%)** |


## UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|377.7 ms|50187.5000|10000.0000|211.26 MB|
| **New** |	|	| **352.1 ms (-7%)** | **52312.5000 (+4%)** | **875.0000 (-91%)** | **210.48 MB (0%)** |